# apps/iana/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api-iana
  version: 0.1.0
  description: |
    API HTTP (JSON), só leitura, dos registros de numeração da IANA coletados
    pelo `collector-iana` do BadBlock (tabelas `iana_*`): a quem cada faixa de
    ASN e cada bloco IP (os /8 do IPv4 e os blocos do IPv6 unicast) foi
    entregue (RIR, status, WHOIS, RDAP), os registros de uso especial
    (special-purpose) de IPv4, IPv6 e ASN, que sustentam a regra de **bogon**,
    e o bootstrap RDAP (RFC 9224). Fonte: https://www.iana.org/numbers.

    Versões: os servidores sem `/v1` servem a versão atual (hoje v1); os
    servidores com `/v1` fixam a v1 (`/iana/v1/asn/61610`), com o mesmo
    conteúdo, a mesma chave de cache e o mesmo ETag. Cada rota aparece uma vez
    em `paths`, relativa ao servidor. Saúde (`/health`, `/status`, `/ping`) e
    este manifesto (`/openapi.yaml`) ficam fora do versionamento: sobrescrevem
    `servers` só com os servidores sem `/v1`.

    Todo `GET` também aceita `HEAD`. `OPTIONS` em qualquer caminho responde o
    preflight de CORS (`204`, `Access-Control-Allow-Methods: GET, HEAD,
    OPTIONS`). `/iana` redireciona (`301`) para `/iana/`; qualquer outro
    caminho ou método responde `404` com `code` `not_found`.

    Em todas as respostas: `X-Content-Type-Options: nosniff`,
    `Referrer-Policy: no-referrer`, `Access-Control-Allow-Origin` (padrão
    `*`), `Access-Control-Expose-Headers: ETag, X-Cache, X-Dataset-Version` e
...

...
  license:
    name: MIT
    identifier: MIT
...
servers:
...
paths:
...
  /asn/{asn}:
    get:
      operationId: getASN
      tags: [dados]
      summary: Consulta um ASN
      description: |
        A faixa de `as-numbers-1`/`-2` que contém o ASN (`block`), as faixas
        de `special-purpose-as-numbers` que o contêm (`special`, em ordem de
        início) e a entrada do bootstrap RDAP (`asn.json`) que o contém
        (`rdap`). Nunca responde 404 para um ASN válido: o que a IANA não diz
        sai `null`/`[]` (`block` só é `null` com dados incompletos; faixas
        reservadas e não alocadas não têm `rdap`).
      parameters:
        - $ref: '#/components/parameters/ASN'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: O que a IANA diz sobre o ASN.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPublic'
            X-Cache:
              $ref: '#/components/headers/XCache'
            X-Dataset-Version:
              $ref: '#/components/headers/XDatasetVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASNResponse'
              examples:
                lacnic:
                  summary: GET /asn/61610
                  value:
                    asn: 61610
                    block:
                      start: 61440
                      end: 61951
                      description: Assigned by LACNIC
                      registry: lacnic
                      whois: whois.lacnic.net
                      rdap_urls: [https://rdap.lacnic.net/rdap/]
                      reference: null
                      registration_date: '2013-06-11'
                    special: []
                    rdap:
                      resource: 61440-61951
                      registry: lacnic
                      urls: [https://rdap.lacnic.net/rdap/]
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
                as_trans:
                  summary: GET /asn/AS23456
                  value:
                    asn: 23456
                    block:
                      start: 23456
                      end: 23456
                      description: AS_TRANS
                      registry: null
                      whois: null
                      rdap_urls: []
                      reference: '[RFC6793]'
                      registration_date: null
                    special:
                      - start: 23456
                        end: 23456
                        reason: AS_TRANS; reserved by [RFC6793]
                        reference: '[RFC6793]'
                    rdap: null
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
  /ip/{ip}:
    get:
      operationId: getIP
      tags: [dados]
      summary: Consulta um IP
      description: |
        O bloco mais específico de `ipv4-address-space` (/8) ou
        `ipv6-unicast-address-assignments` que contém o IP (`block`, `null`
        fora deles), **todas** as entradas dos special-purpose registries
        que o contêm, da mais específica para a menos, inclusive as
        encerradas (`special`), se é **bogon** (`bogon`) e a entrada mais
        específica do bootstrap RDAP que o contém (`rdap`). Nunca responde
        404 para um IP válido.

        Regra de bogon (`store.PrefixLookup.Bogon`), avaliada em ordem:

        1. Entre as entradas de `special` **em vigor** (`termination_date`
           nulo), a mais específica decide se tiver `globally_reachable`
           preenchido: `false` = bogon, `true` = não bogon (ex.:
           `192.0.0.9/32`, `true`, vale dentro de `192.0.0.0/24`, `false`).
           Se ela tem `globally_reachable` nulo (a IANA publica `N/A`:
           TEREDO `2001::/32`, 6to4 `2002::/16`), vale a regra 2. Entradas
           encerradas não contam.
        2. Bogon se nenhum bloco da IANA que cruza a consulta tem status
           diferente de `RESERVED`: a consulta está toda em blocos
           `RESERVED` (0/8, 10/8, 127/8, 224–255...) ou fora de qualquer
           bloco (IPv6 que a IANA não entregou: `ff00::/8`, `4000::/3`...).

        Marca o espaço reservado (bogons clássicos), não o "fullbogon" dos
        RIRs.
      parameters:
        - $ref: '#/components/parameters/IP'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: O que a IANA diz sobre o IP.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPublic'
            X-Cache:
              $ref: '#/components/headers/XCache'
            X-Dataset-Version:
              $ref: '#/components/headers/XDatasetVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPResponse'
              examples:
                pcp_anycast:
                  summary: GET /ip/192.0.0.9 (não bogon)
                  value:
                    ip: 192.0.0.9
                    block:
                      prefix: 192.0.0.0/8
                      designation: Administered by ARIN
                      registry: arin
                      status: LEGACY
                      whois: whois.arin.net
                      rdap_urls:
                        - https://rdap.arin.net/registry
                        - http://rdap.arin.net/registry
                      allocation_date: '1993-05'
                      note: '[10][11]'
                    special:
                      - prefix: 192.0.0.9/32
                        name: Port Control Protocol Anycast
                        rfc: '[RFC7723]'
                        allocation_date: '2015-10'
                        termination_date: null
                        source: true
                        destination: true
                        forwardable: true
                        globally_reachable: true
                        reserved_by_protocol: false
                      - prefix: 192.0.0.0/24
                        name: IETF Protocol Assignments
                        rfc: '[RFC6890], Section 2.1'
                        allocation_date: '2010-01'
                        termination_date: null
                        source: false
                        destination: false
                        forwardable: false
                        globally_reachable: false
                        reserved_by_protocol: false
                    bogon: false
                    rdap:
                      resource: 192.0.0.0/8
                      registry: arin
                      urls:
                        - https://rdap.arin.net/registry/
                        - http://rdap.arin.net/registry/
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
                private_use:
                  summary: GET /ip/10.0.0.1 (bogon)
                  value:
                    ip: 10.0.0.1
                    block:
                      prefix: 10.0.0.0/8
                      designation: IANA - Private Use
                      registry: null
                      status: RESERVED
                      whois: null
                      rdap_urls: []
                      allocation_date: '1995-06'
                      note: '[4]'
                    special:
                      - prefix: 10.0.0.0/8
                        name: Private-Use
                        rfc: '[RFC1918]'
                        allocation_date: '1996-02'
                        termination_date: null
                        source: true
                        destination: true
                        forwardable: true
                        globally_reachable: false
                        reserved_by_protocol: false
                    bogon: true
                    rdap: null
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
  /prefix/{ip}/{len}:
    get:
      operationId: getPrefix
      tags: [dados]
      summary: Consulta um prefixo
      description: |
        Como `/ip`, para um prefixo. Os bits de host são zerados
        (`/prefix/10.1.2.3/16` consulta `10.1.0.0/16`). `block` e `rdap` são
        os mais específicos que **contêm** o prefixo (`null` para um prefixo
        maior que eles, ex.: `8.0.0.0/7`); `special` lista as entradas que
        contêm o prefixo pedido (as que estão dentro dele não entram). `bogon`
        segue a regra do `/ip`, aplicada ao prefixo inteiro (`224.0.0.0/4`
        é bogon: só cruza blocos `RESERVED`; `8.0.0.0/7` não: cruza
        `8.0.0.0/8` `LEGACY`). Nunca responde 404 para um prefixo válido.
      parameters:
        - $ref: '#/components/parameters/PrefixAddress'
        - $ref: '#/components/parameters/PrefixLength'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: O que a IANA diz sobre o prefixo.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPublic'
            X-Cache:
              $ref: '#/components/headers/XCache'
            X-Dataset-Version:
              $ref: '#/components/headers/XDatasetVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrefixResponse'
              example:
                query: 2001:db8::/48
                block:
                  prefix: 2001:c00::/23
                  designation: APNIC
                  registry: apnic
                  status: ALLOCATED
                  whois: whois.apnic.net
                  rdap_urls: [https://rdap.apnic.net/]
                  allocation_date: '2002-05-02'
                  note: 2001:db8::/32 is reserved for Documentation [RFC3849]. See [IPv6 Special-Purpose Address Space] for details.
                special:
                  - prefix: 2001:db8::/32
                    name: Documentation
                    rfc: '[RFC3849]'
                    allocation_date: '2004-07'
                    termination_date: null
                    source: false
                    destination: false
                    forwardable: false
                    globally_reachable: false
                    reserved_by_protocol: false
                bogon: true
                rdap:
                  resource: 2001:c00::/23
                  registry: apnic
                  urls: [https://rdap.apnic.net/]
                dataset:
                  version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                  updated_at: '2026-09-29T00:49:05Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
  /asns:
    get:
      operationId: getASNs
      tags: [dados]
      summary: Todas as faixas de ASN
      description: |
        As faixas de `as-numbers-1`/`-2` (173, ~35 KB), em ordem numérica, no
        formato de `block` do `/asn`. Sem parâmetros de consulta: a lista vem
        sempre inteira, sem paginação nem filtro (a query string é ignorada).
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Todas as faixas de ASN.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPublic'
            X-Cache:
              $ref: '#/components/headers/XCache'
            X-Dataset-Version:
              $ref: '#/components/headers/XDatasetVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASNBlocksResponse'
              example:
                count: 173
                blocks:
                  - start: 0
                    end: 0
                    description: Reserved
                    registry: null
                    whois: null
                    rdap_urls: []
                    reference: '[RFC7607]'
                    registration_date: null
                  - start: 1
                    end: 1876
                    description: Assigned by ARIN
                    registry: arin
                    whois: whois.arin.net
                    rdap_urls:
                      - https://rdap.arin.net/registry
                      - http://rdap.arin.net/registry
                    reference: null
                    registration_date: null
                dataset:
                  version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                  updated_at: '2026-09-29T00:49:05Z'
        '304':
          $ref: '#/components/responses/NotModified'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatasetUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /meta:
    get:
      operationId: getMeta
      tags: [meta]
      summary: Estado dos dados
      description: |
        Último dataset aplicado (os 10 arquivos, na ordem fixa do
        collector-iana) e a linha do `collector-iana` em `jobs`. Sem cache e
        sem ETag; responde também antes da primeira carga, com `dataset` e
        `collector` `null`.
      responses:
        '200':
          description: Estado dos dados.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaResponse'
              examples:
                loaded:
                  summary: Com dados (recorte de 2 dos 10 arquivos)
                  value:
                    app: api-iana
                    version: 0.1.0
                    dataset:
                      version: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
                      updated_at: '2026-09-29T00:49:05Z'
                      sha256: b22ef6b2e76eafd675b5d1a144ad6488720da7c39d56e1d8c09a6cf6401ad06e
                      files:
                        - name: as-numbers-1
                          url: https://www.iana.org/assignments/as-numbers/as-numbers-1.csv
                          sha256: 47f5fe7842029d8077ef749d6b3590293df3252841f1446a188284c5a1839949
                          bytes: 7936
                          rows: 88
                          last_modified: Sat, 19 Sep 2026 00:44:44 GMT
                          publication: null
                        - name: rdap-asn
                          url: https://data.iana.org/rdap/asn.json
                          sha256: 80a0659f933b45130435c0bc7d6143ca2b815b3c5d2ef80c769b985a7821bb65
                          bytes: 4408
                          rows: 159
                          last_modified: Mon, 01 Jun 2026 20:00:01 GMT
                          publication: '2026-06-01T20:00:01Z'
                    collector:
                      app: collector-iana
                      last_sync_at: '2026-09-29T00:49:05Z'
                      last_check_at: '2026-09-29T00:49:05Z'
                      consolidated: false
                empty:
                  summary: Antes da primeira carga
                  value:
                    app: api-iana
                    version: 0.1.0
                    dataset: null
                    collector: null
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
        '504':
          $ref: '#/components/responses/Timeout'
...
  /status:
    servers:
      - url: https://api.badblock.net.br/iana
        description: Produção
      - url: http://127.0.0.1:8107/iana
        description: Desenvolvimento local
    get:
      operationId: getStatus
      tags: [saúde]
      summary: Estado da API
      description: |
        Confere o Postgres, o Valkey e se já há dados. Postgres fora = `error`
        (503); só o Valkey fora = `degraded` (200, a API segue sem cache); sem
        dados ainda = `starting` (200).
      responses:
        '200':
          $ref: '#/components/responses/StatusOK'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/StatusUnavailable'
    post:
      operationId: postStatus
      tags: [saúde]
      summary: Estado da API (POST)
      description: Igual ao `GET /status`; o corpo da requisição é ignorado.
      responses:
        '200':
          $ref: '#/components/responses/StatusOK'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/StatusUnavailable'
...
components:
  parameters:
    ASN:
      name: asn
      in: path
      required: true
      description: Número do AS, de 0 a 4294967295, com ou sem o prefixo `AS` (maiúsculas ou minúsculas).
      schema:
        type: string
        pattern: '^([Aa][Ss])?[0-9]+$'
      example: AS61610
    IP:
      name: ip
      in: path
      required: true
      description: |
        Endereço IPv4 ou IPv6. IPv4 mapeado em IPv6 (`::ffff:10.0.0.1`) vira
        IPv4; endereço com zona (`fe80::1%eth0`) é inválido.
      schema:
        type: string
        anyOf:
          - format: ipv4
          - format: ipv6
      example: 192.0.0.9
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: |
        ETag de uma resposta anterior (fraco ou forte, numa lista, ou `*`). Se
        bater com o ETag atual, a resposta é `304` sem corpo e sem consultar
        nada.
      schema:
        type: string
      example: W/"3db9c89d25843ce7"
  headers:
    ETag:
      description: |
        ETag fraco, `W/"<hash da versão do dataset + consulta normalizada>"`.
        O mesmo com e sem `/v1`; muda quando a versão dos dados muda.
      schema:
        type: string
      example: W/"3db9c89d25843ce7"
    CacheControlPublic:
      description: Cacheável por 5 minutos.
      schema:
        type: string
        const: public, max-age=300
      example: public, max-age=300
...
    XCache:
      description: |
        `HIT` (veio do Valkey, ou `304`), `MISS` (consultou o Postgres e
        guardou no Valkey) ou `BYPASS` (cache desligado).
      schema:
        type: string
        enum: [HIT, MISS, BYPASS]
      example: MISS
    XDatasetVersion:
      description: Versão dos dados usada na resposta (a mesma de `dataset.version`).
      schema:
        type: string
        format: uuid
      example: 01a0eaa2-fd5c-7c24-bc6c-4e3c35d61c53
  responses:
    NotModified:
      description: |
        O `If-None-Match` bateu com o ETag atual: sem corpo e sem consulta ao
        banco (`X-Cache: HIT`).
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Cache-Control:
          $ref: '#/components/headers/CacheControlPublic'
        X-Cache:
          $ref: '#/components/headers/XCache'
        X-Dataset-Version:
          $ref: '#/components/headers/XDatasetVersion'
    BadRequest:
      description: ASN, IP ou tamanho de prefixo inválido (`bad_request`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            asn:
              summary: ASN inválido
              value:
                error:
                  code: bad_request
                  message: 'ASN inválido: use um número de 0 a 4294967295, com ou sem o prefixo AS'
            ip:
              summary: IP inválido
              value:
                error:
                  code: bad_request
                  message: endereço IP inválido
            len:
              summary: Tamanho de prefixo inválido
              value:
                error:
                  code: bad_request
                  message: tamanho de prefixo inválido
...
    DatasetUnavailable:
      description: |
        `dataset_not_ready` (a primeira sincronização do collector-iana ainda
        não terminou) ou `database_unavailable` (falha no Postgres).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            dataset_not_ready:
              summary: Antes da primeira carga
              value:
                error:
                  code: dataset_not_ready
                  message: a primeira sincronização do collector-iana ainda não terminou; tente em alguns minutos
            database_unavailable:
              summary: Falha no Postgres
              value:
                error:
                  code: database_unavailable
                  message: banco de dados indisponível
...
    Timeout:
      description: A consulta passou de `DB_TIMEOUT` (`timeout`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: timeout
              message: a consulta demorou demais
    InternalError:
      description: Panic no tratamento da requisição, registrado no log (`internal_error`).
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: internal_error
              message: erro interno
...
  schemas:
...
    ASNResponse:
      type: object
      required: [asn, block, special, rdap, dataset]
      properties:
        asn:
          type: integer
          format: int64
          minimum: 0
          maximum: 4294967295
          description: ASN consultado, normalizado.
        block:
          description: Faixa que contém o ASN (`null` só com dados incompletos).
          oneOf:
            - $ref: '#/components/schemas/ASNBlock'
            - type: 'null'
        special:
          type: array
          description: Faixas de uso especial que contêm o ASN, em ordem de início (`[]` se nenhuma).
          items:
            $ref: '#/components/schemas/SpecialASN'
        rdap:
          description: Entrada do bootstrap RDAP que contém o ASN, ou `null`.
          oneOf:
            - $ref: '#/components/schemas/RDAPService'
            - type: 'null'
        dataset:
          $ref: '#/components/schemas/Dataset'
...